test(worker-sessions): satisfy review ratchets
